Added top-level component
+ Using ui-view in router instead of components
