+++ app/src/components/all-stores/all-stores.html
@@ -1,0 +1,8 @@
+<p>Stores display</p>
+<section>
+  <span ng-repeat="store in $ctrl.storeList">
+    <h2><a href="#">{{ store.name }}</a></h2>
+      <p>{{ store.city }}, {{ store.state }}</p>
+  </span>
+</section>
+
