+++ app/src/components/all-stores/all-stores.js
@@ -1,0 +1,15 @@
+import template from './stores.html';
+import styles from './stores.scss';
+
+export default {
+  template,
+  controller,
+  bindings: {
+    storeList: '<'
+  }
+};
+
+function controller() {
+  console.log('In stores.js');
+  this.styles = styles;
+}
