--- app/src/components/header/header.html
@@ -1,3 +1,0 @@
-<section ng-class="$ctrl.styles.header">
-  <h1>Pet Store Finder</h1>
-</section>
