--- app/src/components/header/header.js
@@ -1,11 +1,0 @@
-import template from './header.html';
-import styles from './header.scss';
-
-export default {
-  template,
-  controller
-};
-
-function controller() {
-  this.styles = styles;
-}
