+++ app/src/index.html
@@ -5,7 +5,7 @@
     <title>Pet Store</title>
 </head>
 <body>
-    <header></header>
-    <stores></stores>
+    <h1>Pet Store Finder</h1>
+    <ui-view></ui-view>
 </body>
 </html>
